+++ warc-indexer/src/main/java/uk/bl/wa/solr/SolrFields.java
@@ -36,6 +36,9 @@ public interface SolrFields {
 	public static final String SOLR_URL = "url";
 	public static final String SOURCE_FILE = "source_file";
 	public static final String SOURCE_FILE_OFFSET = "source_file_offset";
+        // Intended for building links-graphs. Normalised the same way as SOLR_LINKS
+        public static final String SOLR_URL_NORMALISED = "url_norm";
+	public static final String RESOURCE_NAME = "resourcename"; // type=text_general
 	
 	public static final String SOLR_URL_TYPE = "url_type";
 	public static final String SOLR_URL_TYPE_SLASHPAGE = "slashpage";
@@ -126,5 +129,5 @@ public interface SolrFields {
 	public static final String IMAGE_FACES_COUNT = "image_faces_count";
 	public static final String IMAGE_COLOURS = "image_colours";
 	public static final String IMAGE_DOMINANT_COLOUR = "image_dominant_colour";
-	
+
 }